+++ conjur/api/errors.py
@@ -9,3 +9,7 @@ This module holds Conjur SDK-specific errors for this project
 
 class CertificateHostnameMismatchException(Exception):
     """ Thrown to indicate that a mismatch in the certificate hostname. """
+
+
+class BadInitializationException(Exception):
+    """ Thrown to indicate object that initialized in an invalid way. """