+++ conjur/data_object/credentials_data.py
@@ -6,14 +6,17 @@ CredentialsData module
 This module represents the DTO that holds credential data
 """
 
+
 # pylint: disable=too-few-public-methods
 class CredentialsData:
     """
     Used for setting user input data to login to Conjur
     """
+
+
     def __init__(self, machine: str = None, login: str = None, password: str = None):
         self.machine = machine
-        self.login = login
+        self.login = login # TODO rename to username
         self.password = password
 
     @classmethod
@@ -31,4 +34,5 @@ class CredentialsData:
         """
         Method for comparing resources by their values and not by reference
         """
-        return self.machine == other.machine and self.login == other.login and self.password == other.password
+        return self.machine == other.machine and self.login == other.login and self.password == \
+            other.password